+++ Runtime/Providers/Components/RigidbodyProvider.cs
@@ -1,0 +1,35 @@
+using System;
+using UnityEngine;
+
+namespace VaporEvents
+{
+    public class RigidbodyProvider : ProviderData
+    {
+        private Rigidbody cached;
+        private Func<Rigidbody> OnRequestRaised;
+
+        public void Subscribe(Func<Rigidbody> listener)
+        {
+            OnRequestRaised += listener;
+        }
+
+        public void Unsubscribe()
+        {
+            OnRequestRaised = null;
+        }
+
+        public Rigidbody Request()
+        {
+            if (cached != null)
+            {
+                return cached;
+            }
+            else
+            {
+                Debug.Assert(OnRequestRaised != null, $"RigidbodyProvider was requested before any events were subscribed.");
+                cached = OnRequestRaised.Invoke();
+            }
+            return cached;
+        }
+    }
+}